+++ src/modules/swap/swap.c
@@ -25,9 +25,9 @@ void ffPrintSwap(FFSwapOptions* options)
     FF_STRBUF_AUTO_DESTROY totalPretty = ffStrbufCreate();
     ffParseSize(storage.bytesTotal, &totalPretty);
 
-    uint8_t percentage = storage.bytesTotal == 0
+    double percentage = storage.bytesTotal == 0
         ? 0
-        : (uint8_t) (((long double) storage.bytesUsed / (long double) storage.bytesTotal) * 100.0);
+        : (double) storage.bytesUsed / (double) storage.bytesTotal * 100.0;
 
     if(options->moduleArgs.outputFormat.length == 0)
     {
@@ -37,7 +37,7 @@ void ffPrintSwap(FFSwapOptions* options)
         {
             if(instance.config.percentType & FF_PERCENTAGE_TYPE_BAR_BIT)
             {
-                ffAppendPercentBar(&str, 0, 0, 5, 8);
+                ffAppendPercentBar(&str, 0, 0, 50, 80);
                 ffStrbufAppendC(&str, ' ');
             }
             if(!(instance.config.percentType & FF_PERCENTAGE_TYPE_HIDE_OTHERS_BIT))
@@ -47,7 +47,7 @@ void ffPrintSwap(FFSwapOptions* options)
         {
             if(instance.config.percentType & FF_PERCENTAGE_TYPE_BAR_BIT)
             {
-                ffAppendPercentBar(&str, percentage, 0, 5, 8);
+                ffAppendPercentBar(&str, percentage, 0, 50, 80);
                 ffStrbufAppendC(&str, ' ');
             }
 
@@ -55,7 +55,7 @@ void ffPrintSwap(FFSwapOptions* options)
                 ffStrbufAppendF(&str, "%s / %s ", usedPretty.chars, totalPretty.chars);
 
             if(instance.config.percentType & FF_PERCENTAGE_TYPE_NUM_BIT)
-                ffAppendPercentNum(&str, (uint8_t) percentage, 50, 80, str.length > 0);
+                ffAppendPercentNum(&str, percentage, 50, 80, str.length > 0);
         }
 
         ffStrbufTrimRight(&str, ' ');