+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Changes:
 
 Features:
 * Add `--key-width` for aligning the left edge of values, supported both for global `--key-width` and specific module `--module-key-width`
-* Add `--bar-char-elapsed`, `--bar-char-total` and `--bar-border` options
+* Add `--bar-char-elapsed`, `--bar-char-total`, `--bar-width` and `--bar-border` options
 
 Bugfixes:
 * Fix label detection. Use `--disk-key 'Disk ({2})'` to display it (Disk, Linux)